pyproject.toml
[project]
name = "expensetrackermcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.3",
    "psycopg2-binary>=2.9.9",
    "bcrypt>=4.0.0",
    "pyjwt>=2.8.0",
    "python-dotenv>=1.0.0",
]
